requirements.txt
matplotlib
numpy
pandas
mplfinance
TA-Lib
python-binance
requests